complete CRUD API requests
add the deleteUser and searchUser controllers. the former was quite
simple to implement, the latter needed some logic involving a search
filter helper.

I defined the search controller and within it called the buildFilterQuery() located in the helpers. This filter takes the entire JSON query input and loops through each value, crafting a unique query each time depending on the parameters inputted.

This greatly helps modularizing the codebase by preventing hardcoding of the values within the controller and also preventing complex logic in it.

With this, it returns an array of exact matches depending on input. this
will be used as for search functionality in dwella.